+++ src/plugin-info.ts
@@ -1,4 +1,5 @@
-/* File will be overwritten by bin/release.sh! */export const PLUGIN_INFO = {
-  "pluginVersion": "1.2.0",
-  "pluginReleasedAt": "2023-07-23T19:58:08+0200"
+/* File will be overwritten by bin/release.sh! */
+export const PLUGIN_INFO = {
+  "pluginVersion": "2.0.0",
+  "pluginReleasedAt": "2024-03-25T13:57:35+0100"
 }